[util] Fix for intermodule names
This issue was originally discovered by @eunchan.

The top level template code originally used im_netname for determining
the code for dangling nets at both the module instantiation level, as well
as the broadcast tie-off (for modules like kmac_app and edn1).

However, the input to im_netname was different in each case.
For module level instantiation, the input was the module `inter_signal` definition which does not include `req/rsp` as part of the struct name.

For broadcast tie-off, such as kmac_app and edn1, the input was the top level inter_module definition that includes `req/rsp` as part of the struct name already.

This creates a conflict where either the module level didn't contain the suffix or the top level had an extra suffix.

The fix moves the default net name creation directly to when the top level defintion is created and avoids this conflict.

Eventually this same scheme can be ported to im_netname as well without needing to construct on the fly.  But that will be done in a separate PR.

Signed-off-by: Timothy Chen <timothytim@google.com>