[otbn] Flop address bits to select 32-bit word in LSU
Previously the LSU relied upon the controller keeping the incoming
address stable the cycle following a load request so base word loads
could select the correct word from the WLEN data. When the load uses the
call stack as a base address register this is violated. Flopping the
bits for the word select avoids the need for the controller to keep the
address stable.

Fixes #6295

Signed-off-by: Greg Chadwick <gac@lowrisc.org>