[doc] Clarify HMAC and KMAC endianness settings
The intent of this change is to make it clearer the effect that the
`kmac.CFG.msg_endianness` and `hmac.CFG.byte_swap` settings have on
values written to the relevant `MSG_FIFO`. In particular how writes
smaller than a word are affected.

This also modifies the wording for the equivalent settings for the
output registers. In this case just to clarify that each individual
word in the output has its bytes reversed, as opposed to the entire
multi-word value being reversed.

Fixes #6251.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>