[dv/otp_ctrl] improve conditional coverage
This PR improves the conditional coverage by randomly drive the
lc_check_byp_en signal. This bypass signal will bypass the checking for
lc parition's background check if lc_programmed but reset was not
issued. (Otherwise a false alarm will trigger)
Intially I only intend to turn this On after lc_partition is programmed.
But I realized this might lower some conditional coverage.
This PR will randomly drive this signal after dut_init with 25% of
chance of bypassing the check, and 75% chance of keeping the check on.

Signed-off-by: Cindy Chen <chencindy@google.com>