[kmac] Clear the entropy in StRandEdn
This commit addresses the issue #6393 reported by @udinator

As the comment in the code explains, if SW initiates a new hash
operation while the Entropy FSM is refreshing the internal entropy
(waiting for the EDN response), the logic keeps sending same entropy to
the hash logic.

This commit fixed the behavior to clear the entropy valid signal if
following logic consumes the entropy. Then, the logic will wait until
the EDN response comes back.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>