[sw, dif_kmac] Add SHAKE implementation (blocking only)
Add support for SHAKE128 and SHAKE256 to the KMAC DIF. SHAKE is an
XOF function so the output can be larger than the rate (r). When
this occurs additional state needs to be generated using the 'run'
command.

This commit adds some simple SHAKE test cases to the KMAC DIF smoke
test. More comprehensive testing will come later.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>