[otbn] Document data integrity measures
Start the documentation of security features in OTBN with a
documentation of its data integrity measures. This commit adds two new
sections to the documentation:

- A high-level security feature overview, i.e. the "sales pitch."
- A more detailed description of how the features are implemented in the
  Design Details section.

The documentation is a starting point and we will need to add more
features and refine them as we move on with specification and
implementation.

Fixes #4007

Signed-off-by: Philipp Wagner <phw@lowrisc.org>