[prim_secded] Use _i/_o suffix for port names
The SECDED primitives did use `in` and `out` as port names, which isn't
matching our style guide, which mandates the use of `_i` and `_o`
suffixes. Change the definition and all instantiations to use `data_i`
and `data_o` instead.

The motivation to change this wasn't so much me being
style-guide-pedantic, but more the fact that `in` is a keyword in Python
and I couldn't easily use cocotb to write to the signal for a one-off
testbench.

(We do have a circular dependency between Ibex and the OpenTitan
repository in this regard, making it necessary to temporarily patch the
vendored-in Ibex. I'll remove that once this patch lands.)

Signed-off-by: Philipp Wagner <phw@lowrisc.org>