[dv/unr] Fix unr clk rst ports
This PR fixes an UNR clk rst ports naming mismatch.
In previous version, this is a warning so it is not idenfitied as
failure. But the latest version flaged it as an error.
Also this affect some of the UNR analysis - in OTP_CTRL's case, it
affects the FSM coverage.
Thanks Will for helping me debugging that.

Signed-off-by: Cindy Chen <chencindy@google.com>