[otbnsim] Only change loop stack on commit
Before, the loop stack was popped in the step() function, causing the
simulation state to be partially updated before the commit. This is
inconsistent with other state handling, e.g. by the register file
handling the call stack hidden behind x1. Change things so that call
stack pops are only prepared in the step() function, but actually executed
in commit.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>