[doc,hugo] Make blockquote do something
Without this patch, a block quote is rendered exactly the same as any
other paragraph. Add a couple of em's of padding to the left to indent
it.

Don't put e.g. a vertical bar on the left: for the technical
documentation we're writing, a blockquote is much more useful as a
"here's a side note" than as an actual quote.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>