[spi_device] Add mbist_en port
SPI_DEVICE Dual port SRAM receives muxed clock. The expected clock
frequencies are 24MHz (generic mode) and 48MHz (SPI Flash mode).

But the flash mode needs an external clock, SPI_SCK, which is the faster
clock 48MHz.

For the MBIST, it should use the faster clock to test the SPI_DEVICE
internal DPSRAM. This commit uses SCAN_CLOCK as a MBIST clock, as SCAN
clock is 48MHz at this time.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>