[otp_ctrl] Workaround for generated prim depending on generated prim
A generated tech primitive can currently not depend on another generated
tech primitive, unless that dependency is also declared in the core file
of the abstract primitive wrapper.

This leads to issues in certain flows (lint in this example).

This commit introduces a workaround for this problem - see #6604 for
more context.

Signed-off-by: Michael Schaffner <msf@google.com>