[dpi] Fix indexing bug in ecc32_mem_area
This caused a Valgrind error when setting up imem for OTBN. The bug
was that we were trying to read 40 rather than 39 bits of input. The
fix is the std::min() when defining bits_left.

The rest of the patch is just variable name changes and comments to
make it more obvious how everything works (because this took me quite
some time to debug, even though wrote the code in the first place!)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>