[ci] Enable bitstream generation and FPGA testing for the CW310 board
This commit enables the bistream generation and FPGA testing for the
ChipWhisperer CW310 FPGA board in CI. To this end, support for this
board is added to the pytest-based test infrastructure. Finally, the
test configuration is changed to run all tests on the CW310 FPGA board
(including previously disabled OTBN and Entropy Source tests), and only
a small number of tests on the NexysVideo board.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>