[reggen] Make spacing uniform and simplify comments in reg_top
This is a minor "tidy-up" commit. The idea is that this commit makes
whitespace a bit more uniform, and simplifies the rules about what
comments get emitted. To review this patch, look at the changes to
reg_top.sv.tpl (which should hopefully be obviously right); the rest
of the diff is autogenerated.

The following commit will refactor the code in reg_top.sv.tpl itself.
This is less obviously correct, but *that* patch won't have any diffs
in the autogenerated code.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>